+++ apps/bare-expo/metro.config.js
@@ -1,6 +1,8 @@
 // Learn more https://docs.expo.dev/guides/customizing-metro/
 const { getDefaultConfig } = require('expo/metro-config');
+const path = require('path');
 
+const monorepoRoot = path.join(__dirname, '../..');
 const config = getDefaultConfig(__dirname);
 
 config.resolver.assetExts.push(
@@ -18,6 +20,17 @@ config.resolver.blockList = [
   /\/__tests__\//,
   /\/android\/React(Android|Common)\//,
   /\/versioned-react-native\//,
+];
+
+// Minimize the "watched" folders that Metro crawls through to speed up Metro in big monorepos.
+// Note, omitting folders disables Metro from resolving files within these folders
+// This also happens when symlinks falls within these folders, but the real location doesn't.
+config.watchFolders = [
+  __dirname, // Allow Metro to resolve all files within this project
+  path.join(monorepoRoot, 'apps/native-component-list'), // Allow Metro to resolve all files within NCL
+  path.join(monorepoRoot, 'apps/test-suite'), // Allow Metro to resolve all files within test-suite
+  path.join(monorepoRoot, 'packages'), // Allow Metro to resolve all workspace files of the monorepo
+  path.join(monorepoRoot, 'node_modules'), // Allow Metro to resolve "shared" `node_modules` of the monorepo
 ];
 
 module.exports = config;