+++ packages/expo-updates/ios/Tests/UpdatesDatabaseSpec.swift
@@ -9,7 +9,7 @@ class UpdatesDatabaseSpec : ExpoSpec {
   override class func spec() {
     var testDatabaseDir: URL!
     var db: UpdatesDatabase!
-    var manifest: NewManifest!
+    var manifest: ExpoUpdatesManifest!
     var config: UpdatesConfig!
     
     beforeEach {
@@ -27,7 +27,7 @@ class UpdatesDatabaseSpec : ExpoSpec {
         try! db.openDatabase(inDirectory: testDatabaseDir)
       }
       
-      manifest = NewManifest(rawManifestJSON: [
+      manifest = ExpoUpdatesManifest(rawManifestJSON: [
         "runtimeVersion": "1",
         "id": "0eef8214-4833-4089-9dff-b4138a14f196",
         "createdAt": "2020-11-11T00:17:54.797Z",
@@ -50,8 +50,8 @@ class UpdatesDatabaseSpec : ExpoSpec {
     
     describe("foreign keys") {
       it("throws upon foreign key error") {
-        let update = NewUpdate.update(
-          withNewManifest: manifest,
+        let update = ExpoUpdatesUpdate.update(
+          withExpoUpdatesManifest: manifest,
           extensions: [:],
           config: config,
           database: db
@@ -195,13 +195,13 @@ class UpdatesDatabaseSpec : ExpoSpec {
           return asset
         }
         
-        let manifest1 = NewManifest(rawManifestJSON: [
+        let manifest1 = ExpoUpdatesManifest(rawManifestJSON: [
           "runtimeVersion": "1",
           "id": "0eef8214-4833-4089-9dff-b4138a14f196",
           "createdAt": "2020-11-11T00:17:54.797Z",
           "launchAsset": ["url": "https://url.to/bundle1.js", "contentType": "application/javascript"]
         ])
-        let manifest2 = NewManifest(rawManifestJSON: [
+        let manifest2 = ExpoUpdatesManifest(rawManifestJSON: [
           "runtimeVersion": "1",
           "id": "0eef8214-4833-4089-9dff-b4138a14f197",
           "createdAt": "2020-11-11T00:17:55.797Z",
@@ -217,14 +217,14 @@ class UpdatesDatabaseSpec : ExpoSpec {
         asset2.filename = "same-filename.png"
         asset3.filename = "same-filename.png"
         
-        let update1 = NewUpdate.update(
-          withNewManifest: manifest1,
+        let update1 = ExpoUpdatesUpdate.update(
+          withExpoUpdatesManifest: manifest1,
           extensions: [:],
           config: config,
           database: db
         )
-        let update2 = NewUpdate.update(
-          withNewManifest: manifest2,
+        let update2 = ExpoUpdatesUpdate.update(
+          withExpoUpdatesManifest: manifest2,
           extensions: [:],
           config: config,
           database: db