+++ tools/src/commands/PublishDevExpoHomeCommand.ts
@@ -24,7 +24,7 @@ type ExpoCliStateObject = {
   };
 };
 
-const EXPO_HOME_PATH = Directories.getExpoHomeJSDir();
+const EXPO_HOME_PATH = Directories.getExpoGoDir();
 const { EXPO_HOME_DEV_ACCOUNT_USERNAME, EXPO_HOME_DEV_ACCOUNT_PASSWORD } = process.env;
 
 /**
@@ -201,7 +201,7 @@ async function action(options: ActionOptions): Promise<void> {
   console.log(
     chalk.yellow(
       `Finished publishing. Remember to commit changes of ${chalk.magenta(
-        'home/app.json'
+        'apps/expo-go/app.json'
       )} and ${chalk.magenta('dev-home-config.json')}.`
     )
   );